🚚 move DD export code to source file
